requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiomysql==0.2.0
aiosignal==1.4.0
aiosqlite==0.20.0
annotated-types==0.7.0
asyncpg==0.29.0
attrs==25.4.0
certifi==2025.11.12
frozenlist==1.8.0
greenlet==3.2.4
idna==3.11
magic-filter==1.0.12
multidict==6.7.0
propcache==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.2
python-dotenv==1.0.1
SQLAlchemy==2.0.35
typing_extensions==4.15.0
yarl==1.22.0
